backend/svc/api-docs/requirements.txt
aniso8601==9.0.1
blinker==1.7.0
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
Flask==3.0.0
Flask-Cors==4.0.0
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.10
flask-restful-swagger-3==0.5.1
Flask-Testing==0.8.1
gitdb==4.0.11
GitPython==3.1.40
itsdangerous==2.1.2
Jinja2==3.1.2
mando==0.7.1
MarkupSafe==2.1.3
pymongo==4.6.1
pytz==2023.3.post1
radon==6.0.1
six==1.16.0
smmap==5.0.1
Werkzeug==3.0.1